Add StringTable support to SnapshotImageWriter
Added an abstraction for the StringTable backed by a
J9HashTable.  The hashtable allows duplicate strings
to be commonized.

Additionally, entries in the table are linked in the
order they are added to the table so that the offsets
into the data, as written into the segment, can be
maintained.

The StringTable is flexible enough that it can be used
for both the static string table (shstrtab section) and
the dynamic string table.

Unfortunately, using the J9HashTable forces a dependency
on the J9PortLibrary that I'd been hoping to avoid.

Cleanup required to make the string table and the zero
section use a common, and regular, section header writing
api rather than the adhoc one offs in place now

Cleanup also required to remove the "magic" references in
the code - hard coded indexes, etc.

Also fixed the memory leak of program headers by freeing them
in the SnapshotImageWriter destructor.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>