Add .strtab section
Add a mandatory .strtab section to hold the static strings
and the necessary section header for this.

This is preparation for the .symtab section being added.

Once these are enabled, the same tech can be used to add
the .dynstr, .dynsym, and .hash section for the dynamically
required strings.

Example output:
```

DanTest.image:     file format elf64-x86-64

Contents of section .strtab:
 0000 00466f6f 00426172 00                 .Foo.Bar.
```

Signed-off-by: Dan Heidinga <heidinga@redhat.com>