StringTables now know their SectionHeader
Refactor to make the StringTable hold onto the section
header representing it.  This makes the code a little
simpler.

Also refactored initializer lists so that the `, ` comes
first as it makes it easier to refactor the list in the
future.

Removed some unnecessary functions that are only used in
one place - better to inline them so there's a simpler
api to work with.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>