Clarify that _num_section_headers is a full count.
The section headers are written into a section header table
and need to be uniquely assigned an index in that table, regardless
of whether they are part of a Program Header (load time segment) or
not.

This change eagerly assigns them their index so it can be used for
referencing from one section to another without having to wait until
we're ready to write them out into the file.

Also updated the destructor to free the Section Headers that aren't
part of a program header.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>