SymbolTable: split global and local symbols out
readelf wants local symbols first, then global ones.  To make
that work, adding a second pool to the SymbolTable to hold on
to the global/weak symbols separately from the local ones.

Added a example of creating a global symbol to validate that it
works when reading / writing the elf file

Signed-off-by: Dan Heidinga <heidinga@redhat.com>