Rework SymbolTable to provide better abstraction.
* Move to a single SymbolTable that tracks both the global
and the local symbols.  These need to be in the same backing
pool so they can be written to both the .symtab and .dynsym
sections.

* SymbolTableEntries need to have both a .strtab and .dynstr
index for their name fields.  The writer code needs to set
the correct index before writing the field.

* SymbolTable still needs support for the .hash section to be
added to support dynamic symbol loading.

* Eventually will fix up all the APIs to take a SymbolTable::Binding
to determine whether to return the local or global answer rather
than the existing "xxx_local_xxx" & "xxx_global_xxx" apis.

Dynamic Table
* Started support for the .dynamic section / PT_DYNAMIC segment.
* Still not happy with the current api but its early exploration
of how to specify the the DT_* notes that have to be added to
the _Dynamic[].

String Table
* Fixed a bug in the equals function due to returning 0 for equals
when the API wants 1 to be returned in that case.  Found due to
string hash collisions "losing" different strings.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>